updated swagger documentation, and added yaml file for docs
