+++ package.json
@@ -31,6 +31,7 @@
     "express": "^5.1.0",
     "express-rate-limit": "^8.0.1",
     "helmet": "^8.1.0",
+    "js-yaml": "^4.1.0",
     "pg": "^8.16.3",
     "supertest": "^7.1.4",
     "swagger-jsdoc": "^6.2.8",
@@ -43,6 +44,7 @@
     "@eslint/js": "^9.33.0",
     "@types/cors": "^2.8.19",
     "@types/express": "^5.0.3",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^24.3.0",
     "@types/pg": "^8.15.5",
     "@types/swagger-jsdoc": "^6.0.4",
