--- src/middleware/swagger.ts
@@ -1,6606 +1,0 @@
-import swaggerJsdoc from "swagger-jsdoc";
-import swaggerUi from "swagger-ui-express";
-import { Express, Request, Response } from "express";
-import { logger } from "./logger.js";
-
-const options: swaggerJsdoc.Options = {
-  definition: {
-    openapi: "3.0.0",
-    info: {
-      title: "Wallet Savior API",
-      version: "0.1.0",
-      description: "API documentation for Wallet Savior API application",
-    },
-    components: {
-      schemas: {
-        User: {
-          type: 'object',
-          properties: {
-            id: { type: 'string' },
-            name: { type: 'string' },
-            email: { type: 'string' },
-            emailVerified: { type: 'boolean', default: 'Generated at runtime' },
-            image: { type: 'string' },
-            createdAt: { type: 'string', default: 'Generated at runtime' },
-            updatedAt: { type: 'string', default: 'Generated at runtime' }
-          },
-          required: [ 'name', 'email', 'emailVerified', 'createdAt', 'updatedAt' ]
-        },
-        Session: {
-          type: 'object',
-          properties: {
-            id: { type: 'string' },
-            expiresAt: { type: 'string' },
-            token: { type: 'string' },
-            createdAt: { type: 'string' },
-            updatedAt: { type: 'string' },
-            ipAddress: { type: 'string' },
-            userAgent: { type: 'string' },
-            userId: { type: 'string' }
-          },
-          required: [ 'expiresAt', 'token', 'createdAt', 'updatedAt', 'userId' ]
-        },
-        Account: {
-          type: 'object',
-          properties: {
-            id: { type: 'string' },
-            accountId: { type: 'string' },
-            providerId: { type: 'string' },
-            userId: { type: 'string' },
-            accessToken: { type: 'string' },
-            refreshToken: { type: 'string' },
-            idToken: { type: 'string' },
-            accessTokenExpiresAt: { type: 'string' },
-            refreshTokenExpiresAt: { type: 'string' },
-            scope: { type: 'string' },
-            password: { type: 'string' },
-            createdAt: { type: 'string' },
-            updatedAt: { type: 'string' }
-          },
-          required: [
-            'accountId',
-            'providerId',
-            'userId',
-            'createdAt',
-            'updatedAt'
-          ]
-        },
-        Verification: {
-          type: 'object',
-          properties: {
-            id: { type: 'string' },
-            identifier: { type: 'string' },
-            value: { type: 'string' },
-            expiresAt: { type: 'string' },
-            createdAt: { type: 'string', default: 'Generated at runtime' },
-            updatedAt: { type: 'string', default: 'Generated at runtime' }
-          },
-          required: [ 'identifier', 'value', 'expiresAt' ]
-        }
-      },
-      securitySchemes: {
-        apiKeyCookie: {
-          type: 'apiKey',
-          in: 'cookie',
-          name: 'apiKeyCookie',
-          description: 'API Key authentication via cookie'
-        },
-        bearerAuth: {
-          type: 'http',
-          scheme: 'bearer',
-          description: 'Bearer token authentication'
-        }
-      }
-    },
-    security: [ { apiKeyCookie: [], bearerAuth: [] } ],
-    servers: [ { url: 'http://localhost:8080/api/auth' } ],
-    tags: [
-      {
-        name: 'Default',
-        description: 'Default endpoints that are included with Better Auth by default. These endpoints are not part of any plugin.'
-      }
-    ],
-    paths: {
-      '/sign-in/social': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Sign in with a social provider',
-          operationId: 'socialSignIn',
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    callbackURL: { type: 'string', description: undefined },
-                    newUserCallbackURL: { type: 'string', description: undefined },
-                    errorCallbackURL: { type: 'string', description: undefined },
-                    provider: { type: 'string', description: undefined },
-                    disableRedirect: { type: 'string', description: undefined },
-                    idToken: { type: 'string', description: undefined },
-                    scopes: { type: 'string', description: undefined },
-                    requestSignUp: { type: 'string', description: undefined },
-                    loginHint: { type: 'string', description: undefined }
-                  },
-                  required: [ 'provider' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success - Returns either session details or redirect URL',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    description: 'Session response when idToken is provided',
-                    properties: {
-                      redirect: { type: 'boolean', enum: [ false ] },
-                      token: {
-                        type: 'string',
-                        description: 'Session token',
-                        url: { type: 'null', nullable: true },
-                        user: {
-                          type: 'object',
-                          properties: {
-                            id: { type: 'string' },
-                            email: { type: 'string' },
-                            name: { type: 'string', nullable: true },
-                            image: { type: 'string', nullable: true },
-                            emailVerified: { type: 'boolean' },
-                            createdAt: { type: 'string', format: 'date-time' },
-                            updatedAt: { type: 'string', format: 'date-time' }
-                          },
-                          required: [
-                            'id',
-                            'email',
-                            'emailVerified',
-                            'createdAt',
-                            'updatedAt'
-                          ]
-                        }
-                      }
-                    },
-                    required: [ 'redirect', 'token', 'user' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/get-session': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Get the current session',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      session: { '$ref': '#/components/schemas/Session' },
-                      user: { '$ref': '#/components/schemas/User' }
-                    },
-                    required: [ 'session', 'user' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/sign-out': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Sign out the current user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': { schema: { type: 'object', properties: {} } }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { success: { type: 'boolean' } }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/sign-up/email': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Sign up a user using email and password',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    name: {
-                      type: 'string',
-                      description: 'The name of the user'
-                    },
-                    email: {
-                      type: 'string',
-                      description: 'The email of the user'
-                    },
-                    password: {
-                      type: 'string',
-                      description: 'The password of the user'
-                    },
-                    image: {
-                      type: 'string',
-                      description: 'The profile image URL of the user'
-                    },
-                    callbackURL: {
-                      type: 'string',
-                      description: 'The URL to use for email verification callback'
-                    },
-                    rememberMe: {
-                      type: 'boolean',
-                      description: 'If this is false, the session will not be remembered. Default is `true`.'
-                    }
-                  },
-                  required: [ 'name', 'email', 'password' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Successfully created user',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      token: {
-                        type: 'string',
-                        nullable: true,
-                        description: 'Authentication token for the session'
-                      },
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: {
-                            type: 'string',
-                            description: 'The unique identifier of the user'
-                          },
-                          email: {
-                            type: 'string',
-                            format: 'email',
-                            description: 'The email address of the user'
-                          },
-                          name: {
-                            type: 'string',
-                            description: 'The name of the user'
-                          },
-                          image: {
-                            type: 'string',
-                            format: 'uri',
-                            nullable: true,
-                            description: 'The profile image URL of the user'
-                          },
-                          emailVerified: {
-                            type: 'boolean',
-                            description: 'Whether the email has been verified'
-                          },
-                          createdAt: {
-                            type: 'string',
-                            format: 'date-time',
-                            description: 'When the user was created'
-                          },
-                          updatedAt: {
-                            type: 'string',
-                            format: 'date-time',
-                            description: 'When the user was last updated'
-                          }
-                        },
-                        required: [
-                          'id',
-                          'email',
-                          'name',
-                          'emailVerified',
-                          'createdAt',
-                          'updatedAt'
-                        ]
-                      }
-                    },
-                    required: [ 'user' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/sign-in/email': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Sign in with email and password',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    email: { type: 'string', description: 'Email of the user' },
-                    password: {
-                      type: 'string',
-                      description: 'Password of the user'
-                    },
-                    callbackURL: { type: 'string', description: undefined },
-                    rememberMe: { type: 'string', description: undefined }
-                  },
-                  required: [ 'email', 'password' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success - Returns either session details or redirect URL',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    description: 'Session response when idToken is provided',
-                    properties: {
-                      redirect: { type: 'boolean', enum: [ false ] },
-                      token: { type: 'string', description: 'Session token' },
-                      url: { type: 'null', nullable: true },
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: { type: 'string' },
-                          email: { type: 'string' },
-                          name: { type: 'string', nullable: true },
-                          image: { type: 'string', nullable: true },
-                          emailVerified: { type: 'boolean' },
-                          createdAt: { type: 'string', format: 'date-time' },
-                          updatedAt: { type: 'string', format: 'date-time' }
-                        },
-                        required: [
-                          'id',
-                          'email',
-                          'emailVerified',
-                          'createdAt',
-                          'updatedAt'
-                        ]
-                      }
-                    },
-                    required: [ 'redirect', 'token', 'user' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/forget-password': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Send a password reset email to the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    email: {
-                      type: 'string',
-                      description: 'The email address of the user to send a password reset email to'
-                    },
-                    redirectTo: { type: 'string', description: undefined }
-                  },
-                  required: [ 'email' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: { type: 'boolean' },
-                      message: { type: 'string' }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/reset-password': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Reset the password for a user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    newPassword: {
-                      type: 'string',
-                      description: 'The new password to set'
-                    },
-                    token: { type: 'string', description: undefined }
-                  },
-                  required: [ 'newPassword' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { status: { type: 'boolean' } }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/verify-email': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Verify the email of the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [
-            {
-              name: 'token',
-              in: 'query',
-              description: 'The token to verify the email',
-              required: true,
-              schema: { type: 'string' }
-            },
-            {
-              name: 'callbackURL',
-              in: 'query',
-              description: 'The URL to redirect to after email verification',
-              required: false,
-              schema: { type: 'string' }
-            }
-          ],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: { type: 'string', description: 'User ID' },
-                          email: { type: 'string', description: 'User email' },
-                          name: { type: 'string', description: 'User name' },
-                          image: {
-                            type: 'string',
-                            description: 'User image URL'
-                          },
-                          emailVerified: {
-                            type: 'boolean',
-                            description: 'Indicates if the user email is verified'
-                          },
-                          createdAt: {
-                            type: 'string',
-                            description: 'User creation date'
-                          },
-                          updatedAt: {
-                            type: 'string',
-                            description: 'User update date'
-                          }
-                        },
-                        required: [
-                          'id',
-                          'email',
-                          'name',
-                          'image',
-                          'emailVerified',
-                          'createdAt',
-                          'updatedAt'
-                        ]
-                      },
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if the email was verified successfully'
-                      }
-                    },
-                    required: [ 'user', 'status' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/send-verification-email': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Send a verification email to the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    email: {
-                      type: 'string',
-                      description: 'The email to send the verification email to',
-                      example: 'user@example.com'
-                    },
-                    callbackURL: {
-                      type: 'string',
-                      description: 'The URL to use for email verification callback',
-                      example: 'https://example.com/callback',
-                      nullable: true
-                    }
-                  },
-                  required: [ 'email' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if the email was sent successfully',
-                        example: true
-                      }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              description: 'Bad Request',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      message: {
-                        type: 'string',
-                        description: 'Error message',
-                        example: "Verification email isn't enabled"
-                      }
-                    }
-                  }
-                }
-              }
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/change-email': {
-        post: {
-          tags: [ 'Default' ],
-          description: undefined,
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    newEmail: {
-                      type: 'string',
-                      description: 'The new email address to set must be a valid email address'
-                    },
-                    callbackURL: { type: 'string', description: undefined }
-                  },
-                  required: [ 'newEmail' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Email change request processed successfully',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if the request was successful'
-                      },
-                      message: {
-                        type: 'string',
-                        enum: [ 'Email updated', 'Verification email sent' ],
-                        description: 'Status message of the email change process',
-                        nullable: true
-                      }
-                    },
-                    required: [ 'status' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/change-password': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Change the password of the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    newPassword: {
-                      type: 'string',
-                      description: 'The new password to set'
-                    },
-                    currentPassword: {
-                      type: 'string',
-                      description: 'The current password is required'
-                    },
-                    revokeOtherSessions: { type: 'string', description: undefined }
-                  },
-                  required: [ 'newPassword', 'currentPassword' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Password successfully changed',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      token: {
-                        type: 'string',
-                        nullable: true,
-                        description: 'New session token if other sessions were revoked'
-                      },
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: {
-                            type: 'string',
-                            description: 'The unique identifier of the user'
-                          },
-                          email: {
-                            type: 'string',
-                            format: 'email',
-                            description: 'The email address of the user'
-                          },
-                          name: {
-                            type: 'string',
-                            description: 'The name of the user'
-                          },
-                          image: {
-                            type: 'string',
-                            format: 'uri',
-                            nullable: true,
-                            description: 'The profile image URL of the user'
-                          },
-                          emailVerified: {
-                            type: 'boolean',
-                            description: 'Whether the email has been verified'
-                          },
-                          createdAt: {
-                            type: 'string',
-                            format: 'date-time',
-                            description: 'When the user was created'
-                          },
-                          updatedAt: {
-                            type: 'string',
-                            format: 'date-time',
-                            description: 'When the user was last updated'
-                          }
-                        },
-                        required: [
-                          'id',
-                          'email',
-                          'name',
-                          'emailVerified',
-                          'createdAt',
-                          'updatedAt'
-                        ]
-                      }
-                    },
-                    required: [ 'user' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/update-user': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Update the current user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    name: {
-                      type: 'string',
-                      description: 'The name of the user'
-                    },
-                    image: {
-                      type: 'string',
-                      description: 'The image of the user'
-                    }
-                  }
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if the update was successful'
-                      }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/delete-user': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Delete the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    callbackURL: { type: 'string', description: undefined },
-                    password: { type: 'string', description: undefined },
-                    token: { type: 'string', description: undefined }
-                  },
-                  required: []
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'User deletion processed successfully',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      success: {
-                        type: 'boolean',
-                        description: 'Indicates if the operation was successful'
-                      },
-                      message: {
-                        type: 'string',
-                        enum: [ 'User deleted', 'Verification email sent' ],
-                        description: 'Status message of the deletion process'
-                      }
-                    },
-                    required: [ 'success', 'message' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/reset-password/{token}': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Redirects the user to the callback URL with the token',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [
-            {
-              name: 'callbackURL',
-              in: 'query',
-              schema: {
-                type: 'string',
-                description: 'The URL to redirect the user to reset their password'
-              }
-            }
-          ],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { token: { type: 'string' } }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/request-password-reset': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Send a password reset email to the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    email: {
-                      type: 'string',
-                      description: 'The email address of the user to send a password reset email to'
-                    },
-                    redirectTo: { type: 'string', description: undefined }
-                  },
-                  required: [ 'email' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: { type: 'boolean' },
-                      message: { type: 'string' }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/list-sessions': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'List all active sessions for the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'array',
-                    items: { '$ref': '#/components/schemas/Session' }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/revoke-session': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Revoke a single session',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    token: {
-                      type: 'string',
-                      description: 'The token to revoke'
-                    }
-                  },
-                  required: [ 'token' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if the session was revoked successfully'
-                      }
-                    },
-                    required: [ 'status' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/revoke-sessions': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Revoke all sessions for the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': { schema: { type: 'object', properties: {} } }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if all sessions were revoked successfully'
-                      }
-                    },
-                    required: [ 'status' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/revoke-other-sessions': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Revoke all other sessions for the user except the current one',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            content: {
-              'application/json': { schema: { type: 'object', properties: {} } }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      status: {
-                        type: 'boolean',
-                        description: 'Indicates if all other sessions were revoked successfully'
-                      }
-                    },
-                    required: [ 'status' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/link-social': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Link a social account to the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    callbackURL: { type: 'string', description: undefined },
-                    provider: { type: 'string', description: undefined },
-                    idToken: { type: 'string', description: undefined },
-                    requestSignUp: { type: 'string', description: undefined },
-                    scopes: { type: 'string', description: undefined },
-                    errorCallbackURL: { type: 'string', description: undefined }
-                  },
-                  required: [ 'provider' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      url: {
-                        type: 'string',
-                        description: 'The authorization URL to redirect the user to'
-                      },
-                      redirect: {
-                        type: 'boolean',
-                        description: 'Indicates if the user should be redirected to the authorization URL'
-                      },
-                      status: { type: 'boolean' }
-                    },
-                    required: [ 'redirect' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/list-accounts': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'List all accounts linked to the user',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'array',
-                    items: {
-                      type: 'object',
-                      properties: {
-                        id: { type: 'string' },
-                        provider: { type: 'string' },
-                        createdAt: { type: 'string', format: 'date-time' },
-                        updatedAt: { type: 'string', format: 'date-time' }
-                      },
-                      accountId: { type: 'string' },
-                      scopes: { type: 'array', items: { type: 'string' } }
-                    },
-                    required: [
-                      'id',
-                      'provider',
-                      'createdAt',
-                      'updatedAt',
-                      'accountId',
-                      'scopes'
-                    ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/delete-user/callback': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Callback to complete user deletion with verification token',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [
-            {
-              name: 'token',
-              in: 'query',
-              schema: {
-                type: 'string',
-                description: 'The token to verify the deletion request'
-              }
-            },
-            {
-              name: 'callbackURL',
-              in: 'query',
-              schema: { type: 'string', description: undefined }
-            }
-          ],
-          responses: {
-            '200': {
-              description: 'User successfully deleted',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      success: {
-                        type: 'boolean',
-                        description: 'Indicates if the deletion was successful'
-                      },
-                      message: {
-                        type: 'string',
-                        enum: [ 'User deleted' ],
-                        description: 'Confirmation message'
-                      }
-                    },
-                    required: [ 'success', 'message' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/unlink-account': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Unlink an account',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    providerId: { type: 'string', description: undefined },
-                    accountId: { type: 'string', description: undefined }
-                  },
-                  required: [ 'providerId' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { status: { type: 'boolean' } }
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/refresh-token': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Refresh the access token using a refresh token',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    providerId: {
-                      type: 'string',
-                      description: 'The provider ID for the OAuth provider'
-                    },
-                    accountId: { type: 'string', description: undefined },
-                    userId: { type: 'string', description: undefined }
-                  },
-                  required: [ 'providerId' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Access token refreshed successfully',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      tokenType: { type: 'string' },
-                      idToken: { type: 'string' },
-                      accessToken: { type: 'string' },
-                      refreshToken: { type: 'string' },
-                      accessTokenExpiresAt: { type: 'string', format: 'date-time' },
-                      refreshTokenExpiresAt: { type: 'string', format: 'date-time' }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              description: 'Invalid refresh token or provider configuration'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/get-access-token': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Get a valid access token, doing a refresh if needed',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    providerId: {
-                      type: 'string',
-                      description: 'The provider ID for the OAuth provider'
-                    },
-                    accountId: { type: 'string', description: undefined },
-                    userId: { type: 'string', description: undefined }
-                  },
-                  required: [ 'providerId' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'A Valid access token',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      tokenType: { type: 'string' },
-                      idToken: { type: 'string' },
-                      accessToken: { type: 'string' },
-                      refreshToken: { type: 'string' },
-                      accessTokenExpiresAt: { type: 'string', format: 'date-time' },
-                      refreshTokenExpiresAt: { type: 'string', format: 'date-time' }
-                    }
-                  }
-                }
-              }
-            },
-            '400': {
-              description: 'Invalid refresh token or provider configuration'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/account-info': {
-        post: {
-          tags: [ 'Default' ],
-          description: 'Get the account info provided by the provider',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          requestBody: {
-            required: true,
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    accountId: {
-                      type: 'string',
-                      description: 'The provider given account id for which to get the account info'
-                    }
-                  },
-                  required: [ 'accountId' ]
-                }
-              }
-            }
-          },
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: { type: 'string' },
-                          name: { type: 'string' },
-                          email: { type: 'string' },
-                          image: { type: 'string' },
-                          emailVerified: { type: 'boolean' }
-                        },
-                        required: [ 'id', 'emailVerified' ]
-                      },
-                      data: {
-                        type: 'object',
-                        properties: {},
-                        additionalProperties: true
-                      }
-                    },
-                    required: [ 'user', 'data' ],
-                    additionalProperties: false
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/ok': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Check if the API is working',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          responses: {
-            '200': {
-              description: 'API is working',
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: {
-                      ok: {
-                        type: 'boolean',
-                        description: 'Indicates if the API is working'
-                      }
-                    },
-                    required: [ 'ok' ]
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      },
-      '/error': {
-        get: {
-          tags: [ 'Default' ],
-          description: 'Displays an error page',
-          operationId: undefined,
-          security: [ { bearerAuth: [] } ],
-          parameters: [],
-          responses: {
-            '200': {
-              description: 'Success',
-              content: {
-                'text/html': {
-                  schema: {
-                    type: 'string',
-                    description: 'The HTML content of the error page'
-                  }
-                }
-              }
-            },
-            '400': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-            },
-            '401': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } },
-                    required: [ 'message' ]
-                  }
-                }
-              },
-              description: 'Unauthorized. Due to missing or invalid authentication.'
-            },
-            '403': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-            },
-            '404': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Not Found. The requested resource was not found.'
-            },
-            '429': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-            },
-            '500': {
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    properties: { message: { type: 'string' } }
-                  }
-                }
-              },
-              description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-            }
-          }
-        }
-      }
-    }
-  },
-  apis: ["./src/index.ts", "./src/modules/**/*.ts"],
-};
-
-const swaggerSpec = swaggerJsdoc(options);
-
-export const swaggerDocs = (app: Express, port: number) => {
-  app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerSpec));
-
-  app.get("docs.json", (_req: Request, res: Response) => {
-    res.set("Content-Type", "application/json");
-    res.send(swaggerSpec);
-  });
-
-  logger.info(
-    `Swagger documentation available at http://localhost:${port}/docs`
-  );
-};
-
-
-const authDocs = {
-  openapi: '3.1.1',
-
-  security: [ { apiKeyCookie: [], bearerAuth: [] } ],
-  servers: [ { url: 'http://localhost:8080/api/auth' } ],
-  tags: [
-    {
-      name: 'Default',
-      description: 'Default endpoints that are included with Better Auth by default. These endpoints are not part of any plugin.'
-    }
-  ],
-  paths: {
-    '/sign-in/social': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Sign in with a social provider',
-        operationId: 'socialSignIn',
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  callbackURL: { type: 'string', description: undefined },
-                  newUserCallbackURL: { type: 'string', description: undefined },
-                  errorCallbackURL: { type: 'string', description: undefined },
-                  provider: { type: 'string', description: undefined },
-                  disableRedirect: { type: 'string', description: undefined },
-                  idToken: { type: 'string', description: undefined },
-                  scopes: { type: 'string', description: undefined },
-                  requestSignUp: { type: 'string', description: undefined },
-                  loginHint: { type: 'string', description: undefined }
-                },
-                required: [ 'provider' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success - Returns either session details or redirect URL',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  description: 'Session response when idToken is provided',
-                  properties: {
-                    redirect: { type: 'boolean', enum: [ false ] },
-                    token: {
-                      type: 'string',
-                      description: 'Session token',
-                      url: { type: 'null', nullable: true },
-                      user: {
-                        type: 'object',
-                        properties: {
-                          id: { type: 'string' },
-                          email: { type: 'string' },
-                          name: { type: 'string', nullable: true },
-                          image: { type: 'string', nullable: true },
-                          emailVerified: { type: 'boolean' },
-                          createdAt: { type: 'string', format: 'date-time' },
-                          updatedAt: { type: 'string', format: 'date-time' }
-                        },
-                        required: [
-                          'id',
-                          'email',
-                          'emailVerified',
-                          'createdAt',
-                          'updatedAt'
-                        ]
-                      }
-                    }
-                  },
-                  required: [ 'redirect', 'token', 'user' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/get-session': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Get the current session',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    session: { '$ref': '#/components/schemas/Session' },
-                    user: { '$ref': '#/components/schemas/User' }
-                  },
-                  required: [ 'session', 'user' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/sign-out': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Sign out the current user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': { schema: { type: 'object', properties: {} } }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { success: { type: 'boolean' } }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/sign-up/email': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Sign up a user using email and password',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  name: {
-                    type: 'string',
-                    description: 'The name of the user'
-                  },
-                  email: {
-                    type: 'string',
-                    description: 'The email of the user'
-                  },
-                  password: {
-                    type: 'string',
-                    description: 'The password of the user'
-                  },
-                  image: {
-                    type: 'string',
-                    description: 'The profile image URL of the user'
-                  },
-                  callbackURL: {
-                    type: 'string',
-                    description: 'The URL to use for email verification callback'
-                  },
-                  rememberMe: {
-                    type: 'boolean',
-                    description: 'If this is false, the session will not be remembered. Default is `true`.'
-                  }
-                },
-                required: [ 'name', 'email', 'password' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Successfully created user',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    token: {
-                      type: 'string',
-                      nullable: true,
-                      description: 'Authentication token for the session'
-                    },
-                    user: {
-                      type: 'object',
-                      properties: {
-                        id: {
-                          type: 'string',
-                          description: 'The unique identifier of the user'
-                        },
-                        email: {
-                          type: 'string',
-                          format: 'email',
-                          description: 'The email address of the user'
-                        },
-                        name: {
-                          type: 'string',
-                          description: 'The name of the user'
-                        },
-                        image: {
-                          type: 'string',
-                          format: 'uri',
-                          nullable: true,
-                          description: 'The profile image URL of the user'
-                        },
-                        emailVerified: {
-                          type: 'boolean',
-                          description: 'Whether the email has been verified'
-                        },
-                        createdAt: {
-                          type: 'string',
-                          format: 'date-time',
-                          description: 'When the user was created'
-                        },
-                        updatedAt: {
-                          type: 'string',
-                          format: 'date-time',
-                          description: 'When the user was last updated'
-                        }
-                      },
-                      required: [
-                        'id',
-                        'email',
-                        'name',
-                        'emailVerified',
-                        'createdAt',
-                        'updatedAt'
-                      ]
-                    }
-                  },
-                  required: [ 'user' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/sign-in/email': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Sign in with email and password',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  email: { type: 'string', description: 'Email of the user' },
-                  password: {
-                    type: 'string',
-                    description: 'Password of the user'
-                  },
-                  callbackURL: { type: 'string', description: undefined },
-                  rememberMe: { type: 'string', description: undefined }
-                },
-                required: [ 'email', 'password' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success - Returns either session details or redirect URL',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  description: 'Session response when idToken is provided',
-                  properties: {
-                    redirect: { type: 'boolean', enum: [ false ] },
-                    token: { type: 'string', description: 'Session token' },
-                    url: { type: 'null', nullable: true },
-                    user: {
-                      type: 'object',
-                      properties: {
-                        id: { type: 'string' },
-                        email: { type: 'string' },
-                        name: { type: 'string', nullable: true },
-                        image: { type: 'string', nullable: true },
-                        emailVerified: { type: 'boolean' },
-                        createdAt: { type: 'string', format: 'date-time' },
-                        updatedAt: { type: 'string', format: 'date-time' }
-                      },
-                      required: [
-                        'id',
-                        'email',
-                        'emailVerified',
-                        'createdAt',
-                        'updatedAt'
-                      ]
-                    }
-                  },
-                  required: [ 'redirect', 'token', 'user' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/forget-password': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Send a password reset email to the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  email: {
-                    type: 'string',
-                    description: 'The email address of the user to send a password reset email to'
-                  },
-                  redirectTo: { type: 'string', description: undefined }
-                },
-                required: [ 'email' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: { type: 'boolean' },
-                    message: { type: 'string' }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/reset-password': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Reset the password for a user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  newPassword: {
-                    type: 'string',
-                    description: 'The new password to set'
-                  },
-                  token: { type: 'string', description: undefined }
-                },
-                required: [ 'newPassword' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { status: { type: 'boolean' } }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/verify-email': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Verify the email of the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [
-          {
-            name: 'token',
-            in: 'query',
-            description: 'The token to verify the email',
-            required: true,
-            schema: { type: 'string' }
-          },
-          {
-            name: 'callbackURL',
-            in: 'query',
-            description: 'The URL to redirect to after email verification',
-            required: false,
-            schema: { type: 'string' }
-          }
-        ],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    user: {
-                      type: 'object',
-                      properties: {
-                        id: { type: 'string', description: 'User ID' },
-                        email: { type: 'string', description: 'User email' },
-                        name: { type: 'string', description: 'User name' },
-                        image: {
-                          type: 'string',
-                          description: 'User image URL'
-                        },
-                        emailVerified: {
-                          type: 'boolean',
-                          description: 'Indicates if the user email is verified'
-                        },
-                        createdAt: {
-                          type: 'string',
-                          description: 'User creation date'
-                        },
-                        updatedAt: {
-                          type: 'string',
-                          description: 'User update date'
-                        }
-                      },
-                      required: [
-                        'id',
-                        'email',
-                        'name',
-                        'image',
-                        'emailVerified',
-                        'createdAt',
-                        'updatedAt'
-                      ]
-                    },
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if the email was verified successfully'
-                    }
-                  },
-                  required: [ 'user', 'status' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/send-verification-email': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Send a verification email to the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  email: {
-                    type: 'string',
-                    description: 'The email to send the verification email to',
-                    example: 'user@example.com'
-                  },
-                  callbackURL: {
-                    type: 'string',
-                    description: 'The URL to use for email verification callback',
-                    example: 'https://example.com/callback',
-                    nullable: true
-                  }
-                },
-                required: [ 'email' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if the email was sent successfully',
-                      example: true
-                    }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            description: 'Bad Request',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    message: {
-                      type: 'string',
-                      description: 'Error message',
-                      example: "Verification email isn't enabled"
-                    }
-                  }
-                }
-              }
-            }
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/change-email': {
-      post: {
-        tags: [ 'Default' ],
-        description: undefined,
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  newEmail: {
-                    type: 'string',
-                    description: 'The new email address to set must be a valid email address'
-                  },
-                  callbackURL: { type: 'string', description: undefined }
-                },
-                required: [ 'newEmail' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Email change request processed successfully',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if the request was successful'
-                    },
-                    message: {
-                      type: 'string',
-                      enum: [ 'Email updated', 'Verification email sent' ],
-                      description: 'Status message of the email change process',
-                      nullable: true
-                    }
-                  },
-                  required: [ 'status' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/change-password': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Change the password of the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  newPassword: {
-                    type: 'string',
-                    description: 'The new password to set'
-                  },
-                  currentPassword: {
-                    type: 'string',
-                    description: 'The current password is required'
-                  },
-                  revokeOtherSessions: { type: 'string', description: undefined }
-                },
-                required: [ 'newPassword', 'currentPassword' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Password successfully changed',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    token: {
-                      type: 'string',
-                      nullable: true,
-                      description: 'New session token if other sessions were revoked'
-                    },
-                    user: {
-                      type: 'object',
-                      properties: {
-                        id: {
-                          type: 'string',
-                          description: 'The unique identifier of the user'
-                        },
-                        email: {
-                          type: 'string',
-                          format: 'email',
-                          description: 'The email address of the user'
-                        },
-                        name: {
-                          type: 'string',
-                          description: 'The name of the user'
-                        },
-                        image: {
-                          type: 'string',
-                          format: 'uri',
-                          nullable: true,
-                          description: 'The profile image URL of the user'
-                        },
-                        emailVerified: {
-                          type: 'boolean',
-                          description: 'Whether the email has been verified'
-                        },
-                        createdAt: {
-                          type: 'string',
-                          format: 'date-time',
-                          description: 'When the user was created'
-                        },
-                        updatedAt: {
-                          type: 'string',
-                          format: 'date-time',
-                          description: 'When the user was last updated'
-                        }
-                      },
-                      required: [
-                        'id',
-                        'email',
-                        'name',
-                        'emailVerified',
-                        'createdAt',
-                        'updatedAt'
-                      ]
-                    }
-                  },
-                  required: [ 'user' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/update-user': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Update the current user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  name: {
-                    type: 'string',
-                    description: 'The name of the user'
-                  },
-                  image: {
-                    type: 'string',
-                    description: 'The image of the user'
-                  }
-                }
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if the update was successful'
-                    }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/delete-user': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Delete the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  callbackURL: { type: 'string', description: undefined },
-                  password: { type: 'string', description: undefined },
-                  token: { type: 'string', description: undefined }
-                },
-                required: []
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'User deletion processed successfully',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    success: {
-                      type: 'boolean',
-                      description: 'Indicates if the operation was successful'
-                    },
-                    message: {
-                      type: 'string',
-                      enum: [ 'User deleted', 'Verification email sent' ],
-                      description: 'Status message of the deletion process'
-                    }
-                  },
-                  required: [ 'success', 'message' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/reset-password/{token}': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Redirects the user to the callback URL with the token',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [
-          {
-            name: 'callbackURL',
-            in: 'query',
-            schema: {
-              type: 'string',
-              description: 'The URL to redirect the user to reset their password'
-            }
-          }
-        ],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { token: { type: 'string' } }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/request-password-reset': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Send a password reset email to the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  email: {
-                    type: 'string',
-                    description: 'The email address of the user to send a password reset email to'
-                  },
-                  redirectTo: { type: 'string', description: undefined }
-                },
-                required: [ 'email' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: { type: 'boolean' },
-                    message: { type: 'string' }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/list-sessions': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'List all active sessions for the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'array',
-                  items: { '$ref': '#/components/schemas/Session' }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/revoke-session': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Revoke a single session',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  token: {
-                    type: 'string',
-                    description: 'The token to revoke'
-                  }
-                },
-                required: [ 'token' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if the session was revoked successfully'
-                    }
-                  },
-                  required: [ 'status' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/revoke-sessions': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Revoke all sessions for the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': { schema: { type: 'object', properties: {} } }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if all sessions were revoked successfully'
-                    }
-                  },
-                  required: [ 'status' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/revoke-other-sessions': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Revoke all other sessions for the user except the current one',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          content: {
-            'application/json': { schema: { type: 'object', properties: {} } }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    status: {
-                      type: 'boolean',
-                      description: 'Indicates if all other sessions were revoked successfully'
-                    }
-                  },
-                  required: [ 'status' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/link-social': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Link a social account to the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  callbackURL: { type: 'string', description: undefined },
-                  provider: { type: 'string', description: undefined },
-                  idToken: { type: 'string', description: undefined },
-                  requestSignUp: { type: 'string', description: undefined },
-                  scopes: { type: 'string', description: undefined },
-                  errorCallbackURL: { type: 'string', description: undefined }
-                },
-                required: [ 'provider' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    url: {
-                      type: 'string',
-                      description: 'The authorization URL to redirect the user to'
-                    },
-                    redirect: {
-                      type: 'boolean',
-                      description: 'Indicates if the user should be redirected to the authorization URL'
-                    },
-                    status: { type: 'boolean' }
-                  },
-                  required: [ 'redirect' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/list-accounts': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'List all accounts linked to the user',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'array',
-                  items: {
-                    type: 'object',
-                    properties: {
-                      id: { type: 'string' },
-                      provider: { type: 'string' },
-                      createdAt: { type: 'string', format: 'date-time' },
-                      updatedAt: { type: 'string', format: 'date-time' }
-                    },
-                    accountId: { type: 'string' },
-                    scopes: { type: 'array', items: { type: 'string' } }
-                  },
-                  required: [
-                    'id',
-                    'provider',
-                    'createdAt',
-                    'updatedAt',
-                    'accountId',
-                    'scopes'
-                  ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/delete-user/callback': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Callback to complete user deletion with verification token',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [
-          {
-            name: 'token',
-            in: 'query',
-            schema: {
-              type: 'string',
-              description: 'The token to verify the deletion request'
-            }
-          },
-          {
-            name: 'callbackURL',
-            in: 'query',
-            schema: { type: 'string', description: undefined }
-          }
-        ],
-        responses: {
-          '200': {
-            description: 'User successfully deleted',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    success: {
-                      type: 'boolean',
-                      description: 'Indicates if the deletion was successful'
-                    },
-                    message: {
-                      type: 'string',
-                      enum: [ 'User deleted' ],
-                      description: 'Confirmation message'
-                    }
-                  },
-                  required: [ 'success', 'message' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/unlink-account': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Unlink an account',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  providerId: { type: 'string', description: undefined },
-                  accountId: { type: 'string', description: undefined }
-                },
-                required: [ 'providerId' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { status: { type: 'boolean' } }
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/refresh-token': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Refresh the access token using a refresh token',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  providerId: {
-                    type: 'string',
-                    description: 'The provider ID for the OAuth provider'
-                  },
-                  accountId: { type: 'string', description: undefined },
-                  userId: { type: 'string', description: undefined }
-                },
-                required: [ 'providerId' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Access token refreshed successfully',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    tokenType: { type: 'string' },
-                    idToken: { type: 'string' },
-                    accessToken: { type: 'string' },
-                    refreshToken: { type: 'string' },
-                    accessTokenExpiresAt: { type: 'string', format: 'date-time' },
-                    refreshTokenExpiresAt: { type: 'string', format: 'date-time' }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            description: 'Invalid refresh token or provider configuration'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/get-access-token': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Get a valid access token, doing a refresh if needed',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  providerId: {
-                    type: 'string',
-                    description: 'The provider ID for the OAuth provider'
-                  },
-                  accountId: { type: 'string', description: undefined },
-                  userId: { type: 'string', description: undefined }
-                },
-                required: [ 'providerId' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'A Valid access token',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    tokenType: { type: 'string' },
-                    idToken: { type: 'string' },
-                    accessToken: { type: 'string' },
-                    refreshToken: { type: 'string' },
-                    accessTokenExpiresAt: { type: 'string', format: 'date-time' },
-                    refreshTokenExpiresAt: { type: 'string', format: 'date-time' }
-                  }
-                }
-              }
-            }
-          },
-          '400': {
-            description: 'Invalid refresh token or provider configuration'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/account-info': {
-      post: {
-        tags: [ 'Default' ],
-        description: 'Get the account info provided by the provider',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        requestBody: {
-          required: true,
-          content: {
-            'application/json': {
-              schema: {
-                type: 'object',
-                properties: {
-                  accountId: {
-                    type: 'string',
-                    description: 'The provider given account id for which to get the account info'
-                  }
-                },
-                required: [ 'accountId' ]
-              }
-            }
-          }
-        },
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    user: {
-                      type: 'object',
-                      properties: {
-                        id: { type: 'string' },
-                        name: { type: 'string' },
-                        email: { type: 'string' },
-                        image: { type: 'string' },
-                        emailVerified: { type: 'boolean' }
-                      },
-                      required: [ 'id', 'emailVerified' ]
-                    },
-                    data: {
-                      type: 'object',
-                      properties: {},
-                      additionalProperties: true
-                    }
-                  },
-                  required: [ 'user', 'data' ],
-                  additionalProperties: false
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/ok': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Check if the API is working',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        responses: {
-          '200': {
-            description: 'API is working',
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: {
-                    ok: {
-                      type: 'boolean',
-                      description: 'Indicates if the API is working'
-                    }
-                  },
-                  required: [ 'ok' ]
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    },
-    '/error': {
-      get: {
-        tags: [ 'Default' ],
-        description: 'Displays an error page',
-        operationId: undefined,
-        security: [ { bearerAuth: [] } ],
-        parameters: [],
-        responses: {
-          '200': {
-            description: 'Success',
-            content: {
-              'text/html': {
-                schema: {
-                  type: 'string',
-                  description: 'The HTML content of the error page'
-                }
-              }
-            }
-          },
-          '400': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Bad Request. Usually due to missing parameters, or invalid parameters.'
-          },
-          '401': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } },
-                  required: [ 'message' ]
-                }
-              }
-            },
-            description: 'Unauthorized. Due to missing or invalid authentication.'
-          },
-          '403': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Forbidden. You do not have permission to access this resource or to perform this action.'
-          },
-          '404': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Not Found. The requested resource was not found.'
-          },
-          '429': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Too Many Requests. You have exceeded the rate limit. Try again later.'
-          },
-          '500': {
-            content: {
-              'application/json': {
-                schema: {
-                  type: 'object',
-                  properties: { message: { type: 'string' } }
-                }
-              }
-            },
-            description: 'Internal Server Error. This is a problem with the server that you cannot fix.'
-          }
-        }
-      }
-    }
-  }
-}
