+++ src/modules/bankAccount/interfaces/bankAccountRoutes.ts
@@ -8,34 +8,6 @@ import {
 
 const router = Router();
 
-/**
- * @swagger
- * components:
- *   schemas:
- *     BankAccount:
- *       type: object
- *       properties:
- *         name:
- *           type: string
- *           description: The name of the bank account
- */
-
-/**
- * @swagger
- * /api/accounts:
- *   post:
- *     summary: Create a bank account
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             $ref: '#/components/schemas/BankAccount'
- *     responses:
- *       '201':
- *         description: Bank account created successfully
- */
-
 router.post("/", createBankAccountHandler);
 
 router.get("/", getBankAccountsForUser);
