+++ src/utils/logger.ts
@@ -3,8 +3,16 @@ import { createLogger, format, transports } from "winston";
 export const logger = createLogger({
   level: "info",
   format: format.combine(format.timestamp(), format.json()),
+  defaultMeta: {
+    service: "Wallet Savior",
+    buildInfo: {
+      version: "0.1.0",
+      nodeVersion: process.version,
+    },
+  },
   transports: [
     new transports.Console(),
+    new transports.File({ filename: "logs/combined.log" }),
     new transports.File({ filename: "logs/error.log", level: "error" }),
   ],
 });
